Fix response body streaming with Roda/Sinatra
We had a typo when we attempted to nullify the "async.callback"
variable, which caused Roda/Sinatra to use EventMachine as the scheduler
after all, spawning another EM.defer{} inside which the iteration
happens. This caused a problem with Goliath::RackProxy, because it
assumes the iteration will be synchronous, by closing the request right
after iteration is finished. This would close the rack response body
before the iteration has finished instead of after.