Close read end of the Rack input to avoid race conditions
In Goliath::RackProxy::RackInput#write we are pushing data to the queue,
unless is closed. However, the queue might not be closed when evaluating
the condition, but then it could be closed before we attempt to push to
it.

We could add a mutex, but it's better to just perform closing the read
end of the Rack input synchronously inside the even loop. That way queue
both queue pushing and queue closing can happen only inside the same
thread, avoiding the chance for a race condition.