Close queue in deferred thread after all
When there is a long running request and we terminate the server, call
to EM.next_tick will fail with a RuntimError because the reactor was
already terminated, and that doesn't look nice on the console.

So instead we close the queue for reading in the deferred thread after
all, but we rescue ClosedQueueError to account for the race condition
where the queue can be closed between checking whether it's closed and
pushing to the queue.