Send response as soon as the Rack app returned it
Previously if the Rack app returned the response without reading the
response body, goliath-rack_proxy would still wait until the whole
request body was received before sending the response back.

This is not efficient as it means that in case of bad request headers
the client needs to wait until it has sent the whole request body before
it can find out that the headers were wrong.